test: extract legacy state white-box suite
